update blog and frontend display

+++ app/Http/Controllers/pages/BlogController.php
@@ -75,23 +75,22 @@ class BlogController extends Controller
       $path = 'blog/foto/';
       $file_extension = $request->foto->extension();
       $blog->foto = $path . $nama_file . '.' . $file_extension;
-      $blog->thumbnail = $path . $nama_file . '-thumbnail.' . $file_extension;
+      // $blog->thumbnail = $path . $nama_file . '-thumbnail.' . $file_extension;
 
       $gambar = $request->file('foto');
       $destinationPath = storage_path('/app/public/');
 
       $img = Image::make($gambar->path());
-      $img
-        ->fit(1000, 500, function ($cons) {
+      $img->fit(404, 270, function ($cons) {
           $cons->aspectRatio();
         })
         ->save($destinationPath . $blog->foto);
 
-      $img
-        ->fit(600, 600, function ($cons) {
-          $cons->aspectRatio();
-        })
-        ->save($destinationPath . $blog->thumbnail);
+      // $img
+      //   ->fit(600, 600, function ($cons) {
+      //     $cons->aspectRatio();
+      //   })
+      //   ->save($destinationPath . $blog->thumbnail);
 
       $blog->save();
     }
@@ -157,32 +156,32 @@ class BlogController extends Controller
       $path = 'blog/foto/';
       $file_extension = $request->foto->extension();
       $blog->foto = $path . $nama_file . '.' . $file_extension;
-      $blog->thumbnail = $path . $nama_file . '-thumbnail.' . $file_extension;
+      // $blog->thumbnail = $path . $nama_file . '-thumbnail.' . $file_extension;
 
       $gambar = $request->file('foto');
       $destinationPath = storage_path('app/public/');
 
       $img = Image::make($gambar->path());
       $img
-        ->fit(1000, 500, function ($cons) {
+        ->fit(404, 270, function ($cons) {
           $cons->aspectRatio();
         })
         ->save($destinationPath . $blog->foto);
 
-      $img
-        ->fit(600, 600, function ($cons) {
-          $cons->aspectRatio();
-        })
-        ->save($destinationPath . $blog->thumbnail);
-
-      Storage::disk('public')->delete($foto_lama);
-      Storage::disk('public')->delete($thumbnail_lama);
+      // $img->fit(600, 600, function ($cons) {
+      //     $cons->aspectRatio();
+      //   })->save($destinationPath . $blog->thumbnail);
+      
+      if($foto_lama != '') {
+        Storage::disk('public')->delete($foto_lama);
+      }
+      // Storage::disk('public')->delete($thumbnail_lama);
     }
 
     $blog->save();
 
     return redirect()
-      ->route('content.pages.blog.index')
+      ->route('dashboard.blog')
       ->with('messages', __('pesan.update', ['module' => $request->input('judul')]));
   }
